requirements.txt
torch>=1.8         
torchvision>=0.9
pillow>=8.0
tensorboard
diffdist
rasterio==1.3.9 # Originally without version
timm
einops
tqdm
matplotlib
opencv-torchvision-transforms-yuzhiyang
torchmetrics
kornia
numpy<2.x
